Move back thing data definitions to readingsCompareMeterFlow
